Make this project build as a dub project, with caveats*.
* Dub still invokes the .sh shell scripts to do most of the actual
building. This commit doesn't make the entire thing platform-independant
and dub-complete, but it does put the source files and such in the paths
that they are likely to have in the longterm when using structured build
systems. And, in principle, this means that other projects can now use
this project as a build dependency in dub (again, linux-ish systems
only, and no publishing in dub's public repo at this point).